backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.1.0
psycopg2-binary==2.9.9  # PostgreSQL adapter
alembic==1.13.1  # Database migrations

# Google Sheets 
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
gspread==5.12.0

# Drawing generation
matplotlib==3.9.2
reportlab==4.0.7
Pillow==12.0.0
PyPDF2==3.0.1
PyMuPDF==1.26.7
numpy==1.26.0

# Utilities
python-dotenv==1.2.1
pydantic==2.10.5
pydantic-settings==2.7.1
requests==2.32.5

# Development/Testing
pytest==7.4.3
httpx==0.25.2